Make dates in test time zone aware
The datetime passed to the Source could be either aware of non-aware
depending on the user input. This means that the test is not
representative of real application behavior, which always passes time
zone aware datetimes.

If a datetime has no time zone, assume it is in the local time zone and
convert it to UTC.