Take query time zone into account during string formatting
String formatting happened before time zone conversion and the
data_timezone was used for conversion instead of the
data_query_timezone.

Also, the time zone conversion was not working. `replace` does not change
the timestamp, only changes the time zone information associated with
it, while converting requires the actual timestamp to change.

This fixes #64.